removed usage of coconut-interface crate
