--- common/coconut-interface/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "nym-coconut-interface"
-version = "0.1.0"
-edition = "2021"
-description = "Crutch library until there is proper SerDe support for coconut structs"
-license.workspace = true
-
-[dependencies]
-bs58 = "0.4.0"
-getset = "0.1.1"
-serde = { workspace = true, features = ["derive"] }
-thiserror = { workspace = true }
-
-nym-coconut = { path = "../nymcoconut" }
