+++ common/credentials/Cargo.toml
@@ -16,7 +16,7 @@ time = { workspace = true, features = ["serde"] }
 zeroize = { workspace = true }
 
 # I guess temporarily until we get serde support in coconut up and running
-nym-coconut-interface = { path = "../coconut-interface" }
+nym-credentials-interface = { path = "../credentials-interface" }
 nym-crypto = { path = "../crypto", features = ["rand", "asymmetric", "serde"] }
 nym-api-requests = { path = "../../nym-api/nym-api-requests" }
 nym-validator-client = { path = "../client-libs/validator-client", default-features = false }
