+++ common/credentials/src/coconut/bandwidth/freepass.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::coconut::utils::scalar_serde_helper;
-use nym_coconut_interface::{hash_to_scalar, Attribute, PublicAttribute};
+use nym_credentials_interface::{hash_to_scalar, Attribute, PublicAttribute};
 use serde::{Deserialize, Serialize};
 use time::{Duration, OffsetDateTime, Time};
 use zeroize::{Zeroize, ZeroizeOnDrop};
