+++ gateway/src/node/client_handling/bandwidth.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use log::error;
-use nym_coconut_interface::Credential;
 use nym_credentials::coconut::bandwidth::CredentialType;
 use thiserror::Error;
 
@@ -49,12 +48,12 @@ impl Bandwidth {
     }
 }
 
-impl From<Credential> for Bandwidth {
-    fn from(credential: Credential) -> Self {
-        let token_value = credential.voucher_value();
-        let bandwidth_bytes = token_value * nym_network_defaults::BYTES_PER_UTOKEN;
-        Bandwidth {
-            value: bandwidth_bytes,
-        }
-    }
-}
+// impl From<Credential> for Bandwidth {
+//     fn from(credential: Credential) -> Self {
+//         let token_value = credential.voucher_value();
+//         let bandwidth_bytes = token_value * nym_network_defaults::BYTES_PER_UTOKEN;
+//         Bandwidth {
+//             value: bandwidth_bytes,
+//         }
+//     }
+// }
