+++ nym-api/src/coconut/state.rs
@@ -10,8 +10,8 @@ use crate::coconut::storage::CoconutStorageExt;
 use crate::support::storage::NymApiStorage;
 use nym_api_requests::coconut::helpers::issued_credential_plaintext;
 use nym_api_requests::coconut::BlindSignRequestBody;
+use nym_coconut::{BlindedSignature, VerificationKey};
 use nym_coconut_dkg_common::types::EpochId;
-use nym_coconut_interface::{BlindedSignature, VerificationKey};
 use nym_crypto::asymmetric::identity;
 use nym_validator_client::nyxd::{Hash, TxResponse};
 use std::sync::Arc;
